重构动态阈值模块
